updated for spring social 1.1.0M3

+++ src/main/java/org/springframework/social/salesforce/connect/SalesforceOAuth2Template.java
@@ -28,7 +28,7 @@ public class SalesforceOAuth2Template extends OAuth2Template {
 
 
     @Override
-    protected AccessGrant createAccessGrant(String accessToken, String scope, String refreshToken, Integer expiresIn, Map<String, Object> response) {
+    protected AccessGrant createAccessGrant(String accessToken, String scope, String refreshToken, Long expiresIn, Map<String, Object> response) {
         this.instanceUrl = (String) response.get("instance_url");
 
         return super.createAccessGrant(accessToken, scope, refreshToken, expiresIn, response);
